requirements.txt
biopython==1.79
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
cryptography==3.1.1
cycler==0.11.0
fonttools==4.28.3
greenlet==1.1.2
idna==2.10
kiwisolver==1.3.2
matplotlib==3.5.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
numpy==1.21.2
olefile==0.46
packaging==21.3
pandas==1.1.3
Pillow==8.4.0
pip==21.2.2
pycairo==1.20.1
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==3.0.6
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2020.1
PyYAML==6.0
reportlab==3.5.68
requests==2.24.0
scipy==1.7.1
seaborn==0.11.2
setuptools==58.0.4
six==1.15.0
SQLAlchemy==1.4.27
tornado==6.1
tqdm==4.62.3
unicodedata2==13.0.0.post2
urllib3==1.25.11
wheel==0.37.0
win-inet-pton==1.1.0
wincertstore==0.2
